+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes of the Respect\Validation releases are documented in this fi
 - Create "Finite" rule (#397)
 - Create "Identical" rule (#442)
 - Create "Infinite" rule (#397)
+- Create "IntType" rule (#451)
 - Create "KeyNested" rule (#429)
 - Create "KeySet" rule (#374)
 - Create "KeyValue" rule (#441)